Changed batch size max to avoid OOM

+++ Custom_functions/custom_training_func.py
@@ -62,7 +62,7 @@ def get_HPO_params(config, FLAGS, trial, hpt_opt=False):
         else: batch_size_max = int(np.ceil(np.min(FLAGS.available_mem_info)/1250))
         
         lr = trial.suggest_float(name="learning_rate", low=1e-8, high=1e-3)
-        batch_size = trial.suggest_int(name="batch_size", low=1, high=batch_size_max.astype(np.uint8))
+        batch_size = trial.suggest_int(name="batch_size", low=1, high=int(batch_size_max))
         optimizer_used = trial.suggest_categorical(name="optimizer_used", choices=["ADAMW", "SGD"])
         weight_decay = trial.suggest_float(name="weight_decay", low=1e-8, high=2e-2)
         dice_loss_weight = trial.suggest_float(name="dice_loss_weight", low=1, high=25)
